+++ Loam/src/com/almanac/loam/Model/Monster.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.math.Vector2;
 
 public class Monster extends MoveableEntity {
 	
-	public Monster(float SPEED, float width, float height, float x, float y) {
+	public Monster(float SPEED, float width, float height, int x, int y) {
 		super(SPEED, width, height, x, y);
 	}
